Bytes-based upload progress, 100ms throttled callbacks
Change progress tracking from percentage (int 0-100) to cumulative
bytes uploaded (long) across the full stack: upload streams, outbox,
IPC messages, and UI. The UI calculates percentage from bytes/fileSize
locally. Throttle progress callbacks to 100ms in both ProgressStream
and ChunkUploadStream to avoid flooding IPC.

+++ windows/FileNodeClient.Jmap/AccountScopedJmapClient.cs
@@ -479,7 +479,7 @@ public class AccountScopedJmapClient : IJmapClient
     }
 
     public Task<string> UploadBlobChunkedAsync(Stream data, string contentType, long totalSize,
-        Action<int>? onProgress = null, Action<JmapClient.UploadedChunkInfo>? onChunkUploaded = null,
+        Action<long>? onProgress = null, Action<JmapClient.UploadedChunkInfo>? onChunkUploaded = null,
         List<JmapClient.UploadedChunkInfo>? previousChunks = null,
         CancellationToken ct = default)
     {
